[docs] move anchors wrapper to custom Markdown header components
