+++ docs/ui/components/Button/index.ts
@@ -1,2 +1,3 @@
 export { Button } from './Button';
+export type { ButtonProps } from './Button';
 export { ButtonBase } from './ButtonBase';
