+++ docs/ui/components/Text/index.tsx
@@ -4,7 +4,6 @@ import React from 'react';
 
 import { LinkBase, LinkProps } from './Link';
 import { TextComponentProps, TextElement } from './types';
-import { withAnchor } from './withAnchor';
 
 import { durations } from '~/ui/foundations/durations';
 
@@ -55,24 +54,12 @@ const listStyle = css({
   marginLeft: '1.5rem',
 });
 
-export const H1 = withAnchor(
-  createTextComponent(TextElement.H1, css(typography.headers.default.h1))
-);
-export const H2 = withAnchor(
-  createTextComponent(TextElement.H2, css(typography.headers.default.h2))
-);
-export const H3 = withAnchor(
-  createTextComponent(TextElement.H4, css(typography.headers.default.h3))
-);
-export const H4 = withAnchor(
-  createTextComponent(TextElement.H4, css(typography.headers.default.h4))
-);
-export const H5 = withAnchor(
-  createTextComponent(TextElement.H5, css(typography.headers.default.h5))
-);
-export const H6 = withAnchor(
-  createTextComponent(TextElement.H6, css(typography.headers.default.h6))
-);
+export const H1 = createTextComponent(TextElement.H1, css(typography.headers.default.h1));
+export const H2 = createTextComponent(TextElement.H2, css(typography.headers.default.h2));
+export const H3 = createTextComponent(TextElement.H4, css(typography.headers.default.h3));
+export const H4 = createTextComponent(TextElement.H4, css(typography.headers.default.h4));
+export const H5 = createTextComponent(TextElement.H5, css(typography.headers.default.h5));
+export const H6 = createTextComponent(TextElement.H6, css(typography.headers.default.h6));
 export const P = createTextComponent(TextElement.P, css(typography.body.paragraph));
 export const CODE = createTextComponent(TextElement.CODE, css(typography.utility.inlineCode));
 export const LI = createTextComponent(TextElement.LI, css(typography.body.li));
